+++ src/baobab_mtg_products/exceptions/registration/__init__.py
@@ -1,0 +1,17 @@
+"""Exceptions du workflow d'enregistrement et de scan."""
+
+from baobab_mtg_products.exceptions.registration.ambiguous_barcode_resolution_error import (
+    AmbiguousBarcodeResolutionError,
+)
+from baobab_mtg_products.exceptions.registration.invalid_qualification_state_error import (
+    InvalidQualificationStateError,
+)
+from baobab_mtg_products.exceptions.registration.product_not_found_for_workflow_error import (
+    ProductNotFoundForWorkflowError,
+)
+
+__all__ = [
+    "AmbiguousBarcodeResolutionError",
+    "InvalidQualificationStateError",
+    "ProductNotFoundForWorkflowError",
+]